# swagger/swagger-test.json
{
  "swagger": "2.0",
  "info": {
    "title": "N빵 API",
    "description": "SWM-ChocoBread N빵 API 명세서",
    "version": "1.0.0"
  },
  "host": "localhost:5005",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
...
    "/deals/{dealId}/img": {
      "post": {
        "tags": [
          "Deals"
        ],
        "summary": "S3 이미지(Array) 업로드",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/deals/all/{region}": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "지역 전체 거래 GET",
        "description": "",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/deals/create": {
      "post": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 생성",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "link": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "personalPrice": {
                  "example": "any"
                },
                "totalMember": {
                  "example": "any"
                },
                "dealDate": {
                  "example": "any"
                },
                "place": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "region": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/deals/{dealId}": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 세부정보 GET",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 수정",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "personalPrice": {
                  "example": "any"
                },
                "totalMember": {
                  "example": "any"
                },
                "dealDate": {
                  "example": "any"
                },
                "dealPlace": {
                  "example": "any"
                },
                "currentMember": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 삭제",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/deals/{dealId}/join/{userId}": {
      "post": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 참여",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/deals/{dealId}/users/{userId}": {
      "get": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 유저 상태(참여자, 제안자, 참여하지 않음)",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/deals/{dealId}/report": {
      "post": {
        "tags": [
          "Deals"
        ],
        "summary": "거래 신고",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/deals/{dealId}/endRecruit": {
      "post": {
        "tags": [
          "Deals"
        ],
        "summary": "모집 마감하기",
        "description": "",
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "deprecated": true
      }
    },
    "/users/deals/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "마이페이지 거래내역 조회",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  }
}